@@ -4,7 +4,8 @@ version: 2.4.1
 homepage: https://github.com/Polidea/FlutterBleLib
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
+
   # Flutter versions prior to 1.10 did not support the flutter.plugin.platforms map.
   flutter: ">=1.10.0 <2.0.0"
 
@@ -16,7 +17,7 @@ dependencies:
 
 dev_dependencies:
   mockito: ^5.0.3
-  build_runner: ^1.12.2
+  build_runner: ^2.4.11
   pedantic: ^1.11.0
 
   flutter_test:
