@@ -20,7 +20,7 @@ class MyApp extends StatelessWidget {
       title: 'FlutterBleLib example',
       theme: new ThemeData(
         primaryColor: new Color(0xFF0A3D91),
-        accentColor: new Color(0xFFCC0000),
+        canvasColor: new Color(0xFFCC0000),
       ),
       initialRoute: "/",
       routes: <String, WidgetBuilder>{
