@@ -41,18 +41,10 @@ import 'package:flutter_ble_lib/src/_internal.dart';
 import 'package:flutter_ble_lib/src/_managers_for_classes.dart';
 import 'package:flutter_ble_lib/src/util/_transaction_id_generator.dart';
 
-import 'src/_managers_for_classes.dart';
-
-part 'error/ble_error.dart';
-
 part 'ble_manager.dart';
-
 part 'characteristic.dart';
-
 part 'descriptor.dart';
-
+part 'error/ble_error.dart';
 part 'peripheral.dart';
-
 part 'scan_result.dart';
-
 part 'service.dart';
